Add ability to set temp and maxGenLen

+++ server/src/optimodel_server/RequestTypes/__init__.py
@@ -17,5 +17,5 @@ class QueryBody(BaseModel):
     messages: list[ModelMessage]
     modelToUse: str
     speedPriority: SpeedPriority | None = None
-    temperature: int = 0.2
+    temperature: float = 0.2
     maxGenLen: int = 1024
